Update README.md
The ontology course aims to teach the fundamental principles of formally representing concepts and their relationships in a specific domain. Students learn to structure knowledge explicitly, fostering understanding and interoperability among computer systems. The ultimate goal is to train participants in creating ontologies, semantic structures that facilitate the management and exchange of knowledge across diverse domains.